IS-3138: Flyttet useChangeEnhet
