+++ src/components/tildele/oppfolgingsenhet/TildelOppfolgingsenhetModal.tsx
@@ -11,12 +11,12 @@ import {
   Enhet,
   useGetMuligeOppfolgingsenheter,
 } from "@/data/oppfolgingsenhet/useGetMuligeOppfolgingsenheter";
-import { useChangeEnhet } from "@/components/personkort/useChangeEnhet";
 import { useValgtPersonident } from "@/hooks/useValgtBruker";
 import { useBrukerinfoQuery } from "@/data/navbruker/navbrukerQueryHooks";
 import { useVirksomhetQuery } from "@/data/virksomhet/virksomhetQueryHooks";
 import { useVirksomhetsnummerOfLatestOppfolgingstilfelle } from "@/data/oppfolgingstilfelle/person/oppfolgingstilfellePersonQueryHooks";
 import { TildeltNotification } from "@/components/tildele/oppfolgingsenhet/Oppfolgingsenhet";
+import { useChangeEnhet } from "@/data/behandlendeenhet/behandlendeEnhetQueryHooks";
 
 const text = {
   heading: "Endre oppfølgingsenhet",
