+++ src/data/behandlendeenhet/behandlendeEnhetQueryHooks.ts
@@ -1,9 +1,13 @@
 import { SYFOBEHANDLENDEENHET_ROOT } from "@/apiConstants";
-import { get } from "@/api/axios";
-import { useQuery } from "@tanstack/react-query";
+import { get, post } from "@/api/axios";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useValgtPersonident } from "@/hooks/useValgtBruker";
 import { minutesToMillis } from "@/utils/utils";
-import { BehandlendeEnhetResponseDTO } from "@/data/behandlendeenhet/types/BehandlendeEnhetDTOs";
+import {
+  BehandlendeEnhetResponseDTO,
+  TildelOppfolgingsenhetRequestDTO,
+  TildelOppfolgingsenhetResponseDTO,
+} from "@/data/behandlendeenhet/types/BehandlendeEnhetDTOs";
 import { TildeltHistorikkResponseDTO } from "@/hooks/historikk/useTildeltOppfolgingsenhetHistorikk";
 
 const behandlendeenhetGlobalCacheKeyPrefix = "behandlendeenhet";
@@ -46,5 +50,24 @@ export const useTildeltOppfolgingsenhetHistorikkQuery = () => {
     queryKey: behandlendeEnhetQueryKeys.historikk(fnr),
     queryFn: fetchTildelHistorikkHistorikk,
     enabled: !!fnr,
+  });
+};
+
+export const useChangeEnhet = (fnr: string) => {
+  const queryClient = useQueryClient();
+  const path = `${SYFOBEHANDLENDEENHET_ROOT}/oppfolgingsenhet-tildelinger`;
+  const postChangeEnhet = (requestDTO: TildelOppfolgingsenhetRequestDTO) =>
+    post<TildelOppfolgingsenhetResponseDTO>(path, requestDTO);
+  const behandlendeEnhetQueryKey =
+    behandlendeEnhetQueryKeys.behandlendeEnhet(fnr);
+
+  return useMutation({
+    mutationFn: postChangeEnhet,
+    onSuccess: (data: TildelOppfolgingsenhetResponseDTO) => {
+      queryClient.setQueryData(behandlendeEnhetQueryKey, data);
+      return queryClient.invalidateQueries({
+        queryKey: behandlendeenhetGlobalCacheKey(fnr),
+      });
+    },
   });
 };
